Make rubocop play nice
The intention of the rake task is we will add it to TravisCI. I don’t
want builds to fail if there are style problems because they can be
annoyingly subjective - only fail if there are lint problems.
Individuals can run `rubocop` on their local checkout for style issues.

This also resolves warnings from the .rubocop.yml.